Medical AI/new design/requirements_demo.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
openai>=1.40.0
anthropic>=0.34.0
pydantic>=2.7.0
python-dotenv>=1.0.0

